requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
langchain
langchain-experimental
langchain-groq
matplotlib
seaborn
tabulate
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
sqlalchemy
asyncpg
psycopg2-binary
